+++ src/record.h
@@ -14,7 +14,7 @@
 typedef struct record
 {
     unsigned char MAC[6];
-    int INTERFACE;
+    size_t INTERFACE;
     uint16_t VLAN;
     time_t last_seen;
 } record;
@@ -28,10 +28,10 @@ typedef struct bucket
 extern bucket *mac_table[BUCKETS_SIZE];
 
 bucket* bucket_init(record *rec);
-record* initRec(const unsigned char MAC[6], uint16_t VLAN, int INTERFACE);
+record* initRec(const unsigned char MAC[6], uint16_t VLAN, size_t INTERFACE);
 
 record *mac_table_lookup(const unsigned char MAC[6], uint16_t VLAN_ID);
-void mac_table_learn(const unsigned char MAC[6], uint16_t VLAN_ID, int INTERFACE);
+void mac_table_learn(const unsigned char MAC[6], uint16_t VLAN_ID, size_t INTERFACE);
 void mac_table_age(time_t now);
 
 void mac_table_free(void);
